# backend/cmd/api/openapi.yaml
openapi: 3.1.0
info:
    title: University Room Booking API
    description: |
        REST API for the University Room Booking System.
        Allows users to search rooms, view availability timelines, create booking requests,
        and track their status. Administrators approve/reject pending bookings via a dashboard.

        ## Authentication
        - **Access Token**: Short-lived JWT (15 min) stored in an `httpOnly` cookie.
          `Authorization: Bearer <token>` remains supported for API clients.
        - **Refresh Token**: Long-lived token (30 days) stored in an `httpOnly` cookie with rotation on each use.

        ## Pagination
        All list endpoints use cursor-based pagination. The `nextCursor` value from a response
        should be passed as the `cursor` query parameter in the next request.

        ## Time Slots
        All booking times must be in 5-minute increments.
        Working day: 08:00–22:00.
    version: 1.0.0
    contact:
        name: Backend Architecture Team
servers:
    - url: http://localhost:3000/api
      description: Local development
    - url: https://booking.university.edu/api
      description: Production
...
components:
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: Compatibility access token auth for API clients
        AccessCookieAuth:
            type: apiKey
            in: cookie
            name: accessToken
            description: httpOnly access token cookie
        RefreshCookieAuth:
            type: apiKey
            in: cookie
            name: refreshToken
            description: httpOnly refresh token cookie

    # ────────────────────────────────────────────
    # Reusable Schemas
    # ────────────────────────────────────────────
    schemas:
...
        BuildingOption:
            type: object
            required: [code, label]
            properties:
                code:
                    type: string
                label:
                    type: string
        BookingPurposeOption:
            type: object
            required: [code, label]
            properties:
                code:
                    type: string
                label:
                    type: string
        AdminBookingPurpose:
            type: object
            required: [code, labelRu, labelEn, isActive, sortOrder]
            properties:
                code:
                    type: string
                labelRu:
                    type: string
                labelEn:
                    type: string
                isActive:
                    type: boolean
                sortOrder:
                    type: integer
...
paths:
...
    /buildings:
        get:
            operationId: listBuildings
            tags: [Rooms]
            summary: List active buildings
            description: Returns localized building labels using `X-Locale` or `Accept-Language`.
            security:
                - AccessCookieAuth: []
                - BearerAuth: []
            responses:
                "200":
                    description: Building catalog
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [data]
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/BuildingOption"
    /booking-purposes:
        get:
            operationId: listBookingPurposes
            tags: [Bookings]
            summary: List active booking purposes
            description: Returns localized booking purpose labels using `X-Locale` or `Accept-Language`.
            security:
                - AccessCookieAuth: []
                - BearerAuth: []
            responses:
                "200":
                    description: Booking purpose catalog
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [data]
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/BookingPurposeOption"

    # ── Rooms ───────────────────────────────────
...
    /admin/booking-purposes:
        get:
            operationId: listAdminBookingPurposes
            tags: [Admin]
            summary: List all booking purposes
            security:
                - AccessCookieAuth: []
                - BearerAuth: []
            responses:
                "200":
                    description: Booking purposes
                    content:
                        application/json:
                            schema:
                                type: object
                                required: [data]
                                properties:
                                    data:
                                        type: array
                                        items:
                                            $ref: "#/components/schemas/AdminBookingPurpose"
        post:
            operationId: createBookingPurpose
            tags: [Admin]
            summary: Create booking purpose
            security:
                - AccessCookieAuth: []
                - BearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/AdminBookingPurpose"
            responses:
                "201":
                    description: Created booking purpose
    /admin/booking-purposes/{code}:
        put:
            operationId: updateBookingPurpose
            tags: [Admin]
            summary: Update booking purpose labels/status
            parameters:
                - name: code
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Updated booking purpose
        delete:
            operationId: deactivateBookingPurpose
            tags: [Admin]
            summary: Deactivate booking purpose
            parameters:
                - name: code
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "204":
                    description: Deactivated
    /admin/booking-purposes/{code}/reactivate:
        patch:
            operationId: reactivateBookingPurpose
            tags: [Admin]
            summary: Reactivate booking purpose
            parameters:
                - name: code
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: Reactivated booking purpose
    /admin/booking-purposes/{code}/hard:
        delete:
            operationId: hardDeleteBookingPurpose
            tags: [Admin]
            summary: Permanently delete an unused inactive booking purpose
            security:
                - AccessCookieAuth: []
                - BearerAuth: []
            parameters:
                - name: code
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                "204":
                    description: Booking purpose permanently deleted
                "409":
                    description: Booking purpose is used by existing bookings
...